zig: refactor to allow years compile time options
Also simplify day challenges to not need a Day struct or for challenge to have to know about the source file name / path in order to be able to get a resource file.

Now everything is done based on compile time parameters.

Also refactor so I can solve both 2023 and 2024 advent of code days. Yeah its a lot but it works